% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{show_tags}
\alias{show_tags}
\alias{tags}
\alias{load_tag}
\alias{update_tag}
\title{Facilities for loading and updating other packages}
\usage{
show_tags()

load_tag(tag)

update_tag(tag)
}
\arguments{
\item{tag}{A character strong for the tag. Current options
are "text analysis", "feature engineering", and "time series".}
}
\description{
`tidymodels` contains a set of tags for specific topics in 
 modeling and analysis. These can be used to load packages
 in groups.
}
\examples{
show_tags()
}
