% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{tag_show}
\alias{tag_show}
\alias{tags}
\alias{tag_attach}
\alias{tag_update}
\title{Facilities for loading and updating other packages}
\usage{
tag_show()

tag_attach(tag)

tag_update(tag)
}
\arguments{
\item{tag}{A character strong for the tag. Current options
are 'Bayesian analysis', 'deep learning', 'feature engineering',
'parallel processing', 'resampling', 'text analysis',  and
'time series'.}
}
\description{
`tidymodels` contains a set of tags for specific topics in
 modeling and analysis. These can be used to load packages
 in groups.
}
\examples{
tag_show()
}
