% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{tidymodels_deps}
\alias{tidymodels_deps}
\title{List all tidymodels dependencies}
\usage{
tidymodels_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidymodels packages.}
}
\description{
List all tidymodels dependencies
}
