% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidymodels_packages}
\alias{tidymodels_packages}
\title{List all packages in the tidymodels}
\usage{
tidymodels_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tidymodels in the list?}
}
\description{
List all packages in the tidymodels
}
\examples{
tidymodels_packages()
}
