% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels_prefer.R
\name{tidymodels_prefer}
\alias{tidymodels_prefer}
\title{Resolve conflicts between tidymodels packages and others}
\usage{
tidymodels_prefer(quiet = TRUE)
}
\arguments{
\item{quiet}{If `TRUE`, all output will be suppressed}
}
\description{
`tidymodels_prefer()` uses the `conflicted` package to handle common
conflicts with tidymodels and other packages.
}
\details{
The list of conflicts includes about a dozen that are known issues with other
packages.
}
\examples{
tidymodels_prefer(quiet = FALSE)
}
