% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{tidymodels-package}
\alias{tidymodels}
\alias{tidymodels-package}
\title{tidymodels: Easily Install and Load the 'Tidymodels' Packages}
\description{
The tidy modeling "verse" is a collection of packages for 
    modeling and statistical analysis that share the underlying design
    philosophy, grammar, and data structures of the tidyverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidymodels.tidymodels.org}
  \item \url{https://github.com/tidymodels/tidymodels}
  \item Report bugs at \url{https://github.com/tidymodels/tidymodels/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@rstudio.com}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
