% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{tidymodels-package}
\alias{tidymodels}
\alias{tidymodels-package}
\title{tidymodels: Easily Install and Load the 'Tidymodels' Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The tidy modeling "verse" is a collection of packages for modeling and statistical analysis that share the underlying design philosophy, grammar, and data structures of the tidyverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidymodels.tidymodels.org}
  \item \url{https://github.com/tidymodels/tidymodels}
  \item Report bugs at \url{https://github.com/tidymodels/tidymodels/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
