% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels_prefer.R
\name{tidymodels_prefer}
\alias{tidymodels_prefer}
\title{Resolve conflicts between the tidymodels and other packages}
\usage{
tidymodels_prefer(quiet = TRUE)
}
\arguments{
\item{quiet}{If `TRUE`, all output will be suppressed}
}
\description{
`tidymodels_prefer()` uses the `conflicted` package to handle
common conflicts with tidymodels and other packages.
}
\examples{
tidymodels_prefer()
}
