% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodels_prefer.R
\name{tidymodels_prefer}
\alias{tidymodels_prefer}
\title{Resolve conflicts between tidymodels packages and others}
\usage{
tidymodels_prefer(quiet = TRUE)
}
\arguments{
\item{quiet}{If \code{TRUE}, all output will be suppressed}
}
\description{
\code{tidymodels_prefer()} uses the \code{conflicted} package to handle common
conflicts with tidymodels and other packages.
}
\details{
The list of conflicts includes several dozen that are known issues with other
packages.

Note that the \pkg{conflicted} package is used to manage which packages take
precedence. Using \code{tidymodels_prefer()} will turn on general conflict
resolution during the R session.
}
\examples{
tidymodels_prefer(quiet = FALSE)
}
\seealso{
\code{\link[=tidymodels_conflicts]{tidymodels_conflicts()}}
}
