% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{tidymodels_update}
\alias{tidymodels_update}
\title{Update tidymodels packages}
\usage{
tidymodels_update(pkg = "tidymodels", recursive = FALSE, ...)
}
\arguments{
\item{pkg}{A character string for the model being updated.}

\item{recursive}{If \code{TRUE}, will also check all dependencies of
tidymodels packages.}

\item{...}{Extra arguments to pass to \code{\link[utils:install.packages]{utils::install.packages()}}}
}
\value{
Nothing is returned but a message is printed to the
console about which packages (if any) should be installed along
with code to do so.
}
\description{
This will check to see if all tidymodels packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
tidymodels_update()
}
}
